Добавление эмодзи
