+++ packages/client/src/entities/forum/ui/CommentCard/style.module.scss
@@ -14,4 +14,3 @@
     background-color: var(--g-color-base-selection);
   }
 }
-
